first test added to main
